backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
networkx
matplotlib
seaborn
scipy
google-generativeai
python-multipart
python-dotenv
